WIP: Converting and wrapping the Github queries

+++ app/models/github_profile.rb
@@ -17,17 +17,17 @@ class GithubProfile
 
   has_and_belongs_to_many :orgs, class_name: GithubProfile.name.to_s
 
-  ORGANIZATION = "Organization"
-  USER         = "User"
+  ORGANIZATION = 'Organization'
+  USER         = 'User'
   VALID_TYPES  = [ORGANIZATION, USER]
 
-  class << self
-    def for_username(username, since=1.day.ago)
-      find_or_initialize_by(login: username).tap do |profile|
-        if profile.new_record?
-          logger.info "ALERT: No cached profile for user #{username}"
-          profile.refresh!(nil, since)
-        end
+  def self.for_username(username)
+    find_or_initialize_by(login: username).tap do |profile|
+      if profile.new_record?
+        logger.info "ALERT: No cached profile for user #{username}"
+        token = User.with_username(username).try(:github_token)
+        client = GithubOld.new(token)
+        profile.refresh!(client)
       end
     end
   end
@@ -39,12 +39,15 @@ class GithubProfile
         facts << convert_repo_into_fact(repo)
       end
     end
-    GithubRepo.where('contributors.github_id' => github_id, "owner.github_id" => { '$in' => orgs.map(&:github_id) }).all.each do |repo|
+    GithubRepo.where(
+      'contributors.github_id' => github_id,
+      'owner.github_id' => { '$in' => orgs.map(&:github_id) }
+    ).all.each do |repo|
       if repo.original? && repo.significant_contributions?(github_id)
         facts << convert_repo_into_fact(repo, orgrepo = true)
       end
     end
-    facts << Fact.append!("github:#{login}", "github:#{login}", "Joined GitHub", created_at, "https://github.com/#{login}", ['github', 'account-created'])
+    facts << Fact.append!("github:#{login}", "github:#{login}", 'Joined GitHub', created_at, "https://github.com/#{login}", %w(github account-created))
     return facts
   end
 
@@ -69,26 +72,36 @@ class GithubProfile
     Fact.append!("#{repo.html_url}:#{login}", "github:#{login}", repo.name, repo.created_at, repo.html_url, tags, metadata)
   end
 
-  def refresh!(client=nil, since)
-    client   ||= GithubOld.new
+  def refresh!(client)
     username = self.login
 
-    profile   = client.profile(username, since)
-    github_id = profile.delete(:id)
 
-    repos = client.repos_for(username, since).map do |repo|
-      owner, name = repo.owner.login, repo.name
-      GithubRepo.for_owner_and_name(owner, name, client, repo)
+    if profile = client.profile(username)
+      profile.delete(:id)
+    end
+
+    github_id = User.with_username(username).github_id
+
+    begin
+      raw_repos = client.repos_for(username)
+      ap raw_repos
+      raw_repos.map do |repo|
+        owner, name = repo[:owner][:login], repo[:name]
+        GithubRepo.for_owner_and_name(owner, name, client, repo)
+      end
+    rescue => ex
+      ap ex
+      require 'pry'; binding.pry
     end
 
     update_attributes! profile.merge(
-                         github_id: github_id,
-                         followers: client.followers_for(username, since),
-                         following: client.following_for(username, since),
-                         watched:   client.watched_repos_for(username, since),
-                         orgs:      orgs,
-                         repos:     repos.map { |r| { id: r.id, name: r.name } }
-                       )
+      github_id: github_id,
+      followers: client.followers_for(username),
+      following: client.following_for(username),
+      watched:   client.watched_repos_for(username),
+      orgs:      orgs,
+      repos:     repos.map { |r| { id: r.id, name: r.name } }
+    )
   end
 
   def stale?
